requirements.txt
flask==2.2.5
requests==2.31.0
WTForms==3.1.2
Flask_WTF==1.2.1
Werkzeug==2.2.2
flask_sqlalchemy==3.1.1
python-dotenv==1.0.0
psycopg2-binary==2.9.10
flask-login==0.6.2
Bootstrap_Flask==2.2.0
git+https://git@github.com/timonrieger/database-service.git